basic attending functionality
The attend function should be tied to the attend button,  but the server
gives 405 not allowed if the function is called with the button.  It
works fine if it is called by the onclick marker method.  The userID is
hardcoded at the moment,  should use the userID from the facebook login.